amdrequirements.txt
GPUtil==1.4.0
psutil==5.7.2
six==1.15.0
tabulate==0.8.7
pytest-shutil==1.7.0
pyadl==0.1
sysctl==0.2
py-cpuinfo==7.0.0
google_api==0.1.12
gspread==3.6.0
raypy==0.0.2
torch==0.1.2.post2
ai-benchmark==0.1.2
py-cpuinfo==7.0.0
tensorflow_rocm_enhanced==2.3.0